// mobile/task/openapi-typescript-codegen/generated/models/OrderLoanQueryCondition.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type OrderLoanQueryCondition = {
    addTime?: number;
    assetType?: string;
    endTime?: string;
    loanMoney?: number;
    loanTime?: number;
    orderNo?: string;
    pageNum?: number;
    pageSize?: number;
    payTradeNo?: string;
    phoneNo?: string;
    sortField?: string;
    sortOrder?: string;
    startTime?: string;
    status?: string;
    userName?: string;
};